implement slack oauth exchange
This exposes a `/slack/oauth` route that Slack can query to trigger an
OAuth exchange, culminating in the generation of an access token
authorizing the bot to post messages to a workspace.

When just working with one bot app per workspace, the token can be
generated from the app dashboard and copy-pasted into the secrets file.
But if the app is distributed, and a user who isn’t a bot server admin
wants to install the bot into their workspace, the server needs to
handle OAuth.