send slack notifications with web api if token is configured
Adds a way to send notifications without per-channel webhooks. The
sending mechanism will authenticate with a bearer token that is
valid for any channel that the bot has been added to. The channel
is specified in a new `channel` field in the notification payload.

A note on error handling: for web API requests, Slack always
returns a 200 response and communicates the error message in the
response body, so checking for status alone isn't enough.