tests: add cases for local+default branch filter
Test config is updated to include new prefix rules w/ local
branch filters, as well as a global default filter.

[push.branch_filter_local] checks precedence of local filter.
The first two commits match the rule for #backend1, whose empty
local branch filter overrides the global one, so include the
distinct commit only. The last two commits match the rule
for #backend2, whose local branch filter includes the pushed
branch ref, so include both commits.

[push.branch_filter_default] checks fallback to default filter.
The first commit matches rules for #backend1 (by local empty filter)
and #all-push-events (by global filter), of which the former is
chosen by prefix length. Likewise for the second, except it's not
distinct, so only global filter applies. The last two commits
disqualify for the #backend2 filter so also use the global filter.

[push.branch_filter_nomatch] checks when no filter matched.
This push is on a feature branch not matched by any local/global
filter, so it's sent to the default channel.